[#10] Move jicf.py from physics/ into models/.
