+++ src/stanshock/models/jicf.py
